# docs/swagger.yaml
basePath: 192.168.1.45:10010
definitions:
...
  request.LoginReq:
    properties:
      account:
        description: 账号
        type: string
      password:
        description: 密码
        type: string
    type: object
...
  request.RegisterReq:
    properties:
      account:
        description: 账号
        type: string
      account_type:
        description: 账号类型 1:网关管理员 2.应用链管理员 3.跨链通道管理员
        type: integer
      app_chain_name:
        description: 应用链名称
        type: string
      application_chain_ids:
        description: 支持的应用链Id
        type: string
      bandwidth:
        description: 带宽
        type: integer
      channel_grade:
        description: 通道密级
        type: string
      consensus_mechanism:
        description: 共识机制
        type: string
      contact_name:
        description: 联系人姓名
        type: string
      contract_type:
        description: 合约类型
        type: string
      cross_chain_channel_name:
        description: 跨链通道名称
        type: string
      cross_chain_technology:
        description: 支持的跨链技术
        type: string
      gateway_grade:
        description: 网关密级
        type: string
      gateway_id:
        description: 网关Id
        type: integer
      gateway_ids:
        description: 网关Ids
        type: string
      gateway_name:
        description: 网关名称
        type: string
      head_portrait:
        description: 头像
        type: string
      ip:
        description: 网关Ip地址
        type: string
      logo:
        description: 应用链Logo
        type: string
      mail:
        description: 邮箱
        type: string
      mobile_number:
        description: 手机号码
        type: string
      name:
        description: 名称
        type: string
      password:
        description: 密码
        type: string
      province:
        description: 省份
        type: string
      rpc_url:
        description: 应用链RPC
        type: string
      unit:
        description: 用户单位
        type: string
    type: object
...
  response.EmptyResp:
    type: object
...
  response.LoginResp:
    properties:
      identity:
        description: 身份Id(1.云平台管理员 2.网关管理员 3.应用链管理员 4.通道管理员)
        type: integer
      token:
        description: token
        type: string
    type: object
...
host: localhost:10010
info:
  contact:
    email: 624796905@qq.com
    name: 天河国云
    url: http://www.swagger.io/support
  description: 跨链云平台服务接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 跨链云平台服务接口文档
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            $ref: '#/definitions/response.LoginResp'
      summary: 登录
      tags:
      - API.user
  /user/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            $ref: '#/definitions/response.EmptyResp'
      security:
      - LoginToken: []
      summary: 登出
      tags:
      - API.user
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok","data":{}}'
          schema:
            $ref: '#/definitions/response.EmptyResp'
      summary: 注册账号
      tags:
      - API.user
...
swagger: "2.0"